requirements.txt
streamlit==1.31.0
opencv-python<4.10.0
tensorflow==2.15.0
keras==2.15.0
numpy<2.0.0,>=1.23.5
cvzone==1.6.1
mediapipe==0.10.9
pyttsx3==2.99
pyenchant==3.2.2
pillow==10.2.0
